fix: Image s'affichent correctement. Oublie d'un commit dans le get_main_image.
feat: Affiche pour la page d'accueil.
