+++ back/mariadb/init-db.sql
@@ -230,19 +230,19 @@ VALUES (6, 3, 'This is a great movie!', '2022-01-01 00:00:00');
 
 -- jeux de données pour la table images
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (1, 'https://www.imdb.com/title/tt0111161/mediaviewer/rm10105600/', FALSE);
+VALUES (1, 'https://m.media-amazon.com/images/M/MV5BNDE3ODcxYzMtY2YzZC00NmNlLWJiNDMtZDViZWM2MzIxZDYwXkEyXkFqcGdeQXVyNjAwNDUxODI@._V1_.jpg', FALSE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (1, 'https://www.imdb.com/title/tt0111161/mediaviewer/rm10105600/', FALSE);
+VALUES (1, 'https://m.media-amazon.com/images/I/815qtzaP9iL._AC_UF1000,1000_QL80_.jpg', FALSE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (2, 'https://www.imdb.com/title/tt0068646/mediaviewer/rm10105600/', TRUE);
+VALUES (2, 'https://m.media-amazon.com/images/M/MV5BM2MyNjYxNmUtYTAwNi00MTYxLWJmNWYtYzZlODY3ZTk3OTFlXkEyXkFqcGdeQXVyNzkwMjQ5NzM@._V1_FMjpg_UX1000_.jpg', TRUE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (3, 'https://www.imdb.com/title/tt0468569/mediaviewer/rm10105600/', TRUE);
+VALUES (3, 'https://fr.web.img2.acsta.net/medias/nmedia/18/63/97/89/18949761.jpg', TRUE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (4, 'https://www.imdb.com/title/tt0167260/mediaviewer/rm10105600/', TRUE);
+VALUES (4, 'https://m.media-amazon.com/images/M/MV5BN2EyZjM3NzUtNWUzMi00MTgxLWI0NTctMzY4M2VlOTdjZWRiXkEyXkFqcGdeQXVyNDUzOTQ5MjY@._V1_.jpg', TRUE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (5, 'https://www.imdb.com/title/tt0110912/mediaviewer/rm10105600/', TRUE);
+VALUES (5, 'https://www.ecranlarge.com/media/cache/1600x1200/uploads/image/001/121/7p8x4u3o3p1jzmbqny3zaloby3m-861.jpg', TRUE);
 INSERT INTO images (movie_id, image_url, main_image)
-VALUES (6, 'https://www.imdb.com/title/tt0108052/mediaviewer/rm10105600/', TRUE);
+VALUES (6, 'https://m.media-amazon.com/images/I/817sLmprCSL._AC_UF1000,1000_QL80_.jpg', TRUE);
 
 -- TRIGGER
