+++ front/src/components/MovieDetails.vue
@@ -1,0 +1,11 @@
+<script setup>
+
+</script>
+
+<template>
+
+</template>
+
+<style scoped>
+
+</style>
