--- front/src/components/movies_list.vue
@@ -1,128 +1,0 @@
-<template>
-  <div class="movie-card-grid">
-    <div class="movie-card" v-for="movie in movies" :key="movie.id">
-      <img :src="movie.main_image_url" :alt="movie.title" class="movie-image"/>
-      <p> URL : {{ movie.main_image_url }}</p>
-      <div class="movie-info">
-        <h3>{{ movie.title }}</h3>
-        <p>{{ movie.release_date }}</p>
-        <p>{{ movie.description }}</p>
-        <p>Rental Price: {{ movie.daily_rental_price }} €</p>
-        <p>Purchase Price: {{ movie.purchase_price }} €</p>
-        <button @click="showMoreInfo(movie)" class="info-button">Plus d'informations</button>
-        <button @click="editMovie(movie.id)" class="info-button">Modifier</button>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script>
-import axios from "axios";
-
-export default {
-  name: 'MovieCardGrid',
-  props: {
-    movies: Array
-  },
-  async mounted() {
-    console.log("Hey");
-    console.log(this.movies);
-    await this.fetchMainImages();
-    console.log(this.movies);
-  },
-  methods: {
-    showMoreInfo(film) {
-      alert(`Plus d'infos sur : ${film.title}`);
-      // Ici, vous pouvez rediriger l'utilisateur ou ouvrir un modal avec plus d'informations
-    },
-    fetchMainImages() {
-      console.log("Fetching main images");
-      console.log("Here is movies variable : " + this.movies);
-      if (this.movies === undefined) {
-        console.error("Movies is undefined")
-        return;
-      }
-      if (this.movies.length === 0) {
-        console.error("Movies is empty")
-        return;
-      }
-      for (let i = 0; i < this.movies.length; i++) {
-        console.log("Appel d'url : " + "http://localhost:3000/movies/main_image/" + this.movies[i].id)
-        axios.get("http://localhost:3000/movies/main_image/" + this.movies[i].id)
-            .then(response => {
-              this.movies[i].main_image_url = response.data.image_url;
-            })
-            .catch(error => {
-              console.log(error.response.data);
-            });
-      }
-    },
-    editMovie(movieId) {
-      this.$emit('edit-movie', movieId);
-    },
-  },
-};
-</script>
-
-
-<style scoped>
-.movie-card-grid {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr); /* Affiche 4 cartes par ligne */
-  gap: 20px;
-  padding: 20px;
-}
-
-.movie-card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #fff;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-  overflow: hidden;
-  transition: transform 0.3s; /* Ajoute une transition pour un effet au survol */
-}
-
-.movie-card:hover {
-  transform: scale(1.05); /* Effet de zoom au survol */
-}
-
-.movie-image {
-  width: 100%;
-  height: 200px; /* Hauteur fixe pour toutes les images */
-  object-fit: cover; /* Assure que l'image couvre bien l'espace sans être déformée */
-}
-
-.movie-info {
-  padding: 15px;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1; /* Permet aux informations de prendre tout l'espace disponible */
-}
-
-.movie-info > * {
-  margin-bottom: 10px; /* Espacement entre les éléments dans .movie-info */
-}
-
-h3 {
-  margin: 10px 0;
-}
-
-.info-button {
-  padding: 10px 20px;
-  margin-top: 10px;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.info-button:hover {
-  background-color: #0056b3;
-}
-
-</style>
